Factor out GlyphCellView into its own web component
